Fixed the issue with frames not arriving: cv2.setnumthreads
